Upgrade interproscan to version 5.73-104.0
Removed the appls flag and run the tests will the defaul apps, I caught
a few bugs with like this that I would have missed otherwise.

Remove the out format, it output all the 4 flavours. I think this is a
better default that just one output kind, MGnify-ASA for example uses
the TSV and GFF.